CU-weekend-2026/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
PyJWT
alembic
python-multipart
ytmusicapi
yt-dlp
